refactor: renames the types file to types.d.ts
